+++ app/src/main/java/com/mar_iguana/tours/ui/profile/ProfileViewModel.kt
@@ -1,13 +1,13 @@
-package com.mar_iguana.tours.ui.dashboard
+package com.mar_iguana.tours.ui.profile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ProfileViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is profile Fragment"
     }
     val text: LiveData<String> = _text
 }
