+++ app/src/main/java/com/mar_iguana/tours/ui/wishlist/WishlistFragment.kt
@@ -1,4 +1,4 @@
-package com.mar_iguana.tours.ui.dashboard
+package com.mar_iguana.tours.ui.wishlist
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,20 +10,20 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.mar_iguana.tours.R
 
-class DashboardFragment : Fragment() {
+class WishlistFragment : Fragment() {
 
-    private lateinit var dashboardViewModel: DashboardViewModel
+    private lateinit var wishlistViewModel: WishlistViewModel
 
     override fun onCreateView(
             inflater: LayoutInflater,
             container: ViewGroup?,
             savedInstanceState: Bundle?
     ): View? {
-        dashboardViewModel =
-                ViewModelProvider(this).get(DashboardViewModel::class.java)
+        wishlistViewModel =
+                ViewModelProvider(this).get(WishlistViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_wishlist, container, false)
         val textView: TextView = root.findViewById(R.id.text_dashboard)
-        dashboardViewModel.text.observe(viewLifecycleOwner, Observer {
+        wishlistViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
