+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.mar_iguana.tours.ui.dashboard.DashboardFragment"
+        android:name="com.mar_iguana.tours.ui.wishlist.WishlistFragment"
         android:label="@string/title_wishlist"
         tools:layout="@layout/fragment_wishlist" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.mar_iguana.tours.ui.notifications.NotificationsFragment"
+        android:name="com.mar_iguana.tours.ui.profile.ProfileFragment"
         android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
